+++ templates/generateDataLayerFile.js
@@ -5,6 +5,8 @@ const {
   getStateString,
   getPropTypesString,
   newLine,
+  renderDestructuredValues,
+  renderFunctionDeclaration,
   renderString,
 } = require('./utils.js');
 
@@ -16,13 +18,14 @@ function genAsFunction({
   stateInterface,
   hasPropTypes,
   propTypes,
+  props,
 }) {
   const content =
 `import React from 'react';${renderString((hasPropTypes && !renderTs), `${newLine(true)}import PropTypes from 'prop-types';`)}
 import ${fileName} from './${fileName}';
 
 ${propsInterface}${stateInterface}
-function ${fileName}DataLayer(${renderString((hasPropTypes && !renderTs), 'props')}${renderString(renderTs, 'props: Props')}) {
+${renderFunctionDeclaration(`${fileName}DataLayer`, renderTs, hasPropTypes, props)}
   return (
     <${fileName}/>
   );
@@ -64,6 +67,7 @@ export default ${fileName}DataLayer;
 function generateDataLayerFile(fileName, args, params) {
   const renderTs = args.ext === 'ts';
   const { hasProps, propString } = getPropsString(params, 'data');
+  const props = renderDestructuredValues(params.props.data);
   const { hasState, stateString } = getStateString(params, 'data');
   const { hasPropTypes, propTypesString } = getPropTypesString(params, 'data');
   const propsInterface = renderString(renderTs, genInterface('Props', hasProps, propString));
@@ -77,6 +81,7 @@ function generateDataLayerFile(fileName, args, params) {
     stateInterface,
     hasPropTypes,
     propTypes,
+    props,
   };
   if (args.classes) return genAsClass(passedArguments);
   return genAsFunction(passedArguments);