+++ templates/generateReactNativeMainFile.js
@@ -5,6 +5,8 @@ const {
   getStateString,
   getPropTypesString,
   newLine,
+  renderDestructuredValues,
+  renderFunctionDeclaration,
   renderString,
 } = require('./utils.js');
 
@@ -16,6 +18,7 @@ function genAsFunction({
   stateInterface,
   hasPropTypes,
   propTypes,
+  props,
 }) {
   const content =
 `import React from 'react';${renderString((hasPropTypes && !renderTs), `${newLine(true)}import PropTypes from 'prop-types';`)}
@@ -24,7 +27,7 @@ ${args.style
   : "import { StyleSheet, View, Text } from 'react-native';"}
 ${renderString(args.style, `import ${fileName}Style from './${fileName}.style';`)}${newLine(args.style)}
 ${propsInterface}${stateInterface}
-function ${fileName}(${renderString((hasPropTypes && !renderTs), 'props')}${renderString(renderTs, 'props: Props')}) {
+${renderFunctionDeclaration(fileName, renderTs, hasPropTypes, props)}
   return (
     ${args.style
       ? `<View style={${fileName}Style.root}>
@@ -99,6 +102,7 @@ export default ${fileName};
 function generateReactNativeMainFile(fileName, args, params) {
   const renderTs = args.ext === 'ts';
   const { hasProps, propString } = getPropsString(params, 'main');
+  const props = renderDestructuredValues(params.props.main);
   const { hasState, stateString } = getStateString(params, 'main');
   const { hasPropTypes, propTypesString } = getPropTypesString(params, 'main');
   const propsInterface = renderString(renderTs, genInterface('Props', hasProps, propString));
@@ -113,6 +117,7 @@ function generateReactNativeMainFile(fileName, args, params) {
     stateInterface,
     hasPropTypes,
     propTypes,
+    props,
   };
   if (args.classes) return genAsClass(passedArguments);
   return genAsFunction(passedArguments);