+++ templates/generateSingleReactFile.js
@@ -5,6 +5,8 @@ const {
   getStateString,
   getPropTypesString,
   newLine,
+  renderDestructuredValues,
+  renderFunctionDeclaration,
   renderString,
 } = require('./utils.js');
 
@@ -15,6 +17,7 @@ function genAsFunction({
   stateInterface,
   hasPropTypes,
   propTypes,
+  props,
 }) {
   const content =
 `import React from 'react';${renderString((hasPropTypes && !renderTs), `${newLine(true)}import PropTypes from 'prop-types';`)}
@@ -26,7 +29,7 @@ const Container = styled.div({
   },
 });${newLine(renderTs)}
 ${propsInterface}${stateInterface}
-function ${fileName}(${renderString((hasPropTypes && !renderTs), 'props')}${renderString(renderTs, 'props: Props')}) {
+${renderFunctionDeclaration(fileName, renderTs, hasPropTypes, props)}
   return (
     <Container className="root">
       <p>${fileName} component</p>
@@ -77,6 +80,7 @@ export default ${fileName};
 function generateSingleReactFile(fileName, args, params) {
   const renderTs = args.ext === 'ts';
   const { hasProps, propString } = getPropsString(params, 'main');
+  const props = renderDestructuredValues(params.props.main);
   const { hasState, stateString } = getStateString(params, 'main');
   const { hasPropTypes, propTypesString } = getPropTypesString(params, 'main');
   const propsInterface = renderString(renderTs, genInterface('Props', hasProps, propString));
@@ -91,6 +95,7 @@ function generateSingleReactFile(fileName, args, params) {
     stateInterface,
     hasPropTypes,
     propTypes,
+    props,
   };
   if (args.classes) return genAsClass(passedArguments);
   return genAsFunction(passedArguments);